Update CMakeLists.txt
For some reason, installing `soem_shared` target also install the static library that shares the same name but different extension. They are certainly different files, shared and static, and it is the static that's being copied with the install directive.

Keeping the `soem_shared` target to install suffices for now.